+++ frontend/src/metabase/common/components/Table/index.ts
@@ -1,3 +1,2 @@
+export * from "./ClientSortableTable";
 export * from "./Table";
-export * from "./ControlledTable";
-// export * from "./Table.styled";
